LayoutTests/fast/device-orientation/device-orientation-request-permission-denied-expected.txt

CONSOLE MESSAGE: No device orientation events will be fired, reason: Permission to use the API was not yet requested.
Received device orientation & motion access request for security origin "".
CONSOLE MESSAGE: No device orientation events will be fired, reason: Permission to use the API was denied.
Basic testing for DeviceOrientationEvent.requestPermission().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is "denied"
PASS result is "denied"
PASS successfullyParsed is true

TEST COMPLETE
